opengl: separate private API only used in core
