+++ src/gldit/cairo-dock-container.c
@@ -33,7 +33,7 @@
 #include "cairo-dock-log.h"
 #include "cairo-dock-config.h"
 #include "cairo-dock-utils.h"  // cairo_dock_string_is_address
-#include "cairo-dock-opengl.h"
+#include "cairo-dock-opengl-priv.h"
 #include "cairo-dock-animations.h"  // cairo_dock_animation_will_be_visible
 #include "cairo-dock-desktop-manager.h"  // gldi_desktop_get_width
 #include "cairo-dock-menu.h"  // gldi_menu_new
