+++ src/gldit/cairo-dock-core.c
@@ -60,7 +60,7 @@
 #include "cairo-dock-file-manager.h"
 #include "cairo-dock-overlay.h"
 #include "cairo-dock-log.h"
-#include "cairo-dock-opengl.h"
+#include "cairo-dock-opengl-priv.h"
 #include "cairo-dock-core.h"
 
 extern GldiContainer *g_pPrimaryContainer;
